requirements.txt
Django==5.1
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
Pillow==11.0.0
firebase-admin==6.6.0
python-dotenv==1.0.1
drf-spectacular==0.28.0
